+++ lib/graphql/rails/active_reflection/attribute_reflection.rb
@@ -12,17 +12,27 @@ module GraphQL::Rails::ActiveReflection
       @klass = klass
       @field_name = field.name
       @name = field.property || field.name
-      @validators = klass.validators.map { |validator|
-        # Skip if the validator does not apply to this field
-        # Skip if there are :if or :unless options (conditionals purposely unsupported)
-        next if validator.attributes.exclude?(@name.to_sym) || validator.options[:if].present? || validator.options[:unless].present?
-        GraphQL::Rails::ActiveReflection::ValidatorReflection.new(validator)
-      }.compact
       @errors = []
+      get_validators
     end
 
     class << self
       attr_accessor :schema_name
+    end
+
+    protected
+
+    def get_validators
+      @validators ||= klass.validators.map { |validator|
+        # Skip if the validator does not apply to this field
+        # Skip if there are :if or :unless options (conditionals purposely unsupported)
+        next unless include_validator? validator
+        GraphQL::Rails::ActiveReflection::ValidatorReflection.new(validator)
+      }.compact
+    end
+
+    def include_validator? validator
+      validator.attributes.include?(@name.to_sym) && validator.options[:if].nil? && validator.options[:unless].nil?
     end
   end
 end