+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ### Version 0.1.0
 Initial commit. More to come.
+
+### Version 0.2.0
+Added ModelReflection.get_attributes and ModelReflection.include_property?
+Added AttributeReflection.get_validators and ActiveReflection.include_validator?